pyproject.toml
[project]
name = "langclaw"
version = "0.1.0"
requires-python = ">=3.11"
description = "Production-ready multi-channel AI agent framework built on langchain 1.0, langgraph, and deepagents"
dependencies = [
    "langchain>=1.0",
    "langgraph>=0.4",
    "langchain-core>=0.3",
    "deepagents>=0.1",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "apscheduler>=4.0.0a1",
    "typer>=0.12",
    "httpx>=0.27",
    "anyio>=4.0",
    "langgraph-checkpoint-sqlite",
    "langchain-openai>=1.1.10",
    "tenacity>=9.1.4",
    "loguru>=0.7.3",
]

[project.optional-dependencies]
telegram = ["python-telegram-bot>=21"]
discord  = ["discord.py>=2"]
slack    = ["slack-sdk>=3"]
websocket = [
    "websockets>=14",
]
postgres = ["langgraph-checkpoint-postgres", "psycopg[binary,pool]"]
rabbitmq = ["aio-pika>=9"]
kafka    = ["aiokafka>=0.10"]
mcp      = ["langchain-mcp-adapters>=0.1"]
gmail    = ["google-api-python-client>=2.100", "google-auth>=2.20", "google-auth-oauthlib>=1.1"]
search   = ["crawl4ai>=0.4", "langchain-community>=0.3", "tavily-python>=0.3", "duckduckgo-search>=6"]
dev      = ["pytest>=8", "pytest-asyncio>=0.23", "ruff>=0.4"]
all      = ["langclaw[telegram,discord,slack,websocket,postgres,rabbitmq,kafka,mcp,search,gmail]"]

[project.scripts]
langclaw = "langclaw.cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["langclaw"]

[tool.uv]
# apscheduler v4 is still in alpha; allow pre-releases project-wide.
prerelease = "allow"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
